+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## 更新日志
 
+### 2021年4月4日
+
+* 增加CORE_CONFIG_D，方便游戏自定义运行配置，配置文件位置`/data/config`
+* 为了数据安全调整数据库密码配置也保存在自定义配置文件中
+
 ### 2021年3月27日
 
-* 为了数据安全调整数据库密码配置文件位置为`DATA_DIR`下的`DB_CONFIG`
 * 调整CORE_LOGIN_D中 init_new_player 方法为`private void init_new_player(object user, object ob)`
 
 ### 2021年3月10日